requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.0.4
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
outcome==1.2.0
pycparser==2.21
pydantic==1.10.2
PySocks==1.7.1
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.22.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
wsproto==1.2.0
